refactor: MergeStrategy enum, module extraction, and tokio::main async (#72, #73, #75)
- Add MergeStrategy enum (All/Longest/Sparse) replacing bare string literals
- Add named tolerance constants (TOLERANCE_SYNTHETIC_MS, etc.)
- Extract merge functions to src/merge.rs and browser helpers to src/io/browser.rs
- Move config loaders (load_analysis_config, load_tag_rules, etc.) to config.rs
- Convert main.rs to #[tokio::main(flavor=current_thread)] async with LocalSet
- Remove manual tokio::runtime::Builder construction (3 instances)
- Remove VALID_MERGE_STRATEGIES constant (replaced by enum)
- Update pipeline/mod.rs and segmenter.rs to use MergeStrategy enum

Fixes #72, Fixes #73, Fixes #75
