+++ src/main.rs
@@ -3,6 +3,7 @@ mod config;
 mod error;
 mod io;
 mod learning;
+mod merge;
 mod pipeline;
 mod review;
 mod types;
@@ -15,22 +16,28 @@ use tracing::{info, Level};
 use tracing_subscriber::EnvFilter;
 
 use crate::cli::{Cli, Commands};
-use crate::config::{MergeOptions, VALID_MERGE_STRATEGIES};
+use crate::config::{
+    get_calibration_dir, load_analysis_config, load_merge_options, load_tag_rules,
+    tolerance_for_profile,
+};
+use crate::io::browser::open_in_browser;
 use crate::io::json::{read_json, read_timeline, write_json_pretty};
 use crate::learning::calibrator::{apply_calibration_report, run_calibration};
 use crate::learning::profiles::CalibrationProfile;
+use crate::merge::merge_nonvoice_segments;
 use crate::pipeline::prompts::add_prompts;
-use crate::pipeline::tags::{add_tags, TagRules};
+use crate::pipeline::tags::add_tags;
 use crate::pipeline::{benchmark_file, extract_timeline};
 
-fn main() {
-    if let Err(err) = run() {
+#[tokio::main(flavor = "current_thread")]
+async fn main() {
+    if let Err(err) = run().await {
         eprintln!("error: {err:#}");
         std::process::exit(1);
     }
 }
 
-fn run() -> Result<()> {
+async fn run() -> Result<()> {
     init_logging();
     let cli = Cli::parse();
     info!(command = ?cli.command, "timeline command start");
@@ -380,35 +387,39 @@ fn run() -> Result<()> {
                 if let Some(parent) = db_path.parent() {
                     std::fs::create_dir_all(parent)?;
                 }
-                let rt = tokio::runtime::Builder::new_current_thread()
-                    .enable_all()
-                    .build()
-                    .context("failed to create async runtime for learning db")?;
-                let learning_db = rt.block_on(learning::database::LearningDb::new(&db_path))?;
-                for (i, result) in report.segment_results.iter().enumerate() {
-                    let segment = learning::database::VerifiedSegment {
-                        start_ms: result.start_ms as i64,
-                        end_ms: result.end_ms as i64,
-                        confidence: result.original_confidence as f64,
-                        spectral_features: learning::database::SpectralFeatures {
-                            rms: result.spectral_features.rms as f64,
-                            zcr: result.spectral_features.zcr as f64,
-                            spectral_flux: result.spectral_features.spectral_flux as f64,
-                            spectral_flatness: result.spectral_features.spectral_flatness as f64,
-                            spectral_entropy: result.spectral_features.spectral_entropy as f64,
-                            centroid_hz: result.spectral_features.centroid_hz as f64,
-                            low_band_ratio: result.spectral_features.low_band_ratio as f64,
-                            high_band_ratio: result.spectral_features.high_band_ratio as f64,
-                        },
-                        was_false_positive: result.is_suspicious,
-                    };
-                    let segment_id = rt.block_on(learning_db.record_verification(segment))?;
+                let local = tokio::task::LocalSet::new();
+                local
+                    .run_until(async {
+                        let learning_db = learning::database::LearningDb::new(&db_path).await?;
+                        for (i, result) in report.segment_results.iter().enumerate() {
+                            let segment = learning::database::VerifiedSegment {
+                                start_ms: result.start_ms as i64,
+                                end_ms: result.end_ms as i64,
+                                confidence: result.original_confidence as f64,
+                                spectral_features: learning::database::SpectralFeatures {
+                                    rms: result.spectral_features.rms as f64,
+                                    zcr: result.spectral_features.zcr as f64,
+                                    spectral_flux: result.spectral_features.spectral_flux as f64,
+                                    spectral_flatness: result.spectral_features.spectral_flatness
+                                        as f64,
+                                    spectral_entropy: result.spectral_features.spectral_entropy
+                                        as f64,
+                                    centroid_hz: result.spectral_features.centroid_hz as f64,
+                                    low_band_ratio: result.spectral_features.low_band_ratio as f64,
+                                    high_band_ratio: result.spectral_features.high_band_ratio
+                                        as f64,
+                                },
+                                was_false_positive: result.is_suspicious,
+                            };
+                            let segment_id = learning_db.record_verification(segment).await?;
 
-                    // Also store fingerprints
-                    if let Some(fps) = report.segment_fingerprints.get(i) {
-                        rt.block_on(learning_db.record_fingerprints(segment_id, fps))?;
-                    }
-                }
+                            if let Some(fps) = report.segment_fingerprints.get(i) {
+                                learning_db.record_fingerprints(segment_id, fps).await?;
+                            }
+                        }
+                        Ok::<(), anyhow::Error>(())
+                    })
+                    .await?;
                 info!(path = %db_path.display(), "saved learning data to database");
             }
         }
@@ -418,19 +429,22 @@ fn run() -> Result<()> {
             output,
         } => {
             let recommendations_json = if let Some(db_path) = learning_db {
-                let rt = tokio::runtime::Builder::new_current_thread()
-                    .enable_all()
-                    .build()
-                    .context("failed to create async runtime for learning db")?;
-                let learning_db = rt.block_on(learning::database::LearningDb::new(&db_path))?;
-                let recommendations = rt.block_on(learning_db.get_threshold_recommendations())?;
-                rt.block_on(learning_db.record_threshold(
-                    recommendations.suggested_flatness_max,
-                    recommendations.suggested_entropy_min,
-                    recommendations.suggested_centroid_min,
-                    recommendations.suggested_centroid_max,
-                ))?;
-                serde_json::to_vec_pretty(&recommendations)?
+                let local = tokio::task::LocalSet::new();
+                local
+                    .run_until(async {
+                        let learning_db = learning::database::LearningDb::new(&db_path).await?;
+                        let recommendations = learning_db.get_threshold_recommendations().await?;
+                        learning_db
+                            .record_threshold(
+                                recommendations.suggested_flatness_max,
+                                recommendations.suggested_entropy_min,
+                                recommendations.suggested_centroid_min,
+                                recommendations.suggested_centroid_max,
+                            )
+                            .await?;
+                        Ok::<_, anyhow::Error>(serde_json::to_vec_pretty(&recommendations)?)
+                    })
+                    .await?
             } else {
                 let mut state =
                     learning::adaptive_thresholds::load_learning_state(&learning_state)?;
@@ -458,21 +472,22 @@ fn run() -> Result<()> {
             learning_db,
             output,
         } => {
-            let rt = tokio::runtime::Builder::new_current_thread()
-                .enable_all()
-                .build()
-                .context("failed to create async runtime for learning db")?;
-            let db = rt.block_on(learning::database::LearningDb::new(&learning_db))?;
-            let stats = rt.block_on(db.get_statistics())?;
-            let recommendations = rt.block_on(db.get_threshold_recommendations())?;
-            let latest_threshold = rt.block_on(db.get_latest_threshold())?;
+            let local = tokio::task::LocalSet::new();
+            let report = local
+                .run_until(async {
+                    let db = learning::database::LearningDb::new(&learning_db).await?;
+                    let stats = db.get_statistics().await?;
+                    let recommendations = db.get_threshold_recommendations().await?;
+                    let latest_threshold = db.get_latest_threshold().await?;
 
-            let report = serde_json::json!({
-                "learning_db": learning_db,
-                "statistics": stats,
-                "recommendations": recommendations,
-                "latest_threshold": latest_threshold,
-            });
+                    Ok::<_, anyhow::Error>(serde_json::json!({
+                        "learning_db": learning_db,
+                        "statistics": stats,
+                        "recommendations": recommendations,
+                        "latest_threshold": latest_threshold,
+                    }))
+                })
+                .await?;
 
             if let Some(output_path) = output {
                 if let Some(parent) = output_path.parent() {
@@ -591,119 +606,6 @@ fn run() -> Result<()> {
     Ok(())
 }
 
-fn tolerance_for_profile(profile: &str) -> u64 {
-    match profile {
-        "synthetic" => 100,
-        "dataset" => 200,
-        _ => 400,
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn load_analysis_config(
-    config_path: Option<std::path::PathBuf>,
-    threshold_override: Option<f32>,
-    min_speech_override: Option<u32>,
-    min_silence_override: Option<u32>,
-    max_non_voice_override: Option<u32>,
-    vad_engine: String,
-    calibration_profile: Option<std::path::PathBuf>,
-    parallel_features: Option<bool>,
-) -> Result<config::AnalysisConfig> {
-    let threshold_delta = load_calibration_threshold_delta(calibration_profile.as_deref())?;
-    config::AnalysisConfig::from_args(
-        config_path,
-        threshold_override,
-        min_speech_override,
-        min_silence_override,
-        max_non_voice_override,
-        Some(vad_engine),
-        threshold_delta,
-        parallel_features,
-    )
-}
-
-fn load_calibration_threshold_delta(profile_path: Option<&std::path::Path>) -> Result<Option<f32>> {
-    let Some(profile_path) = profile_path else {
-        return Ok(None);
-    };
-    let profile: CalibrationProfile = read_json(profile_path).with_context(|| {
-        format!(
-            "failed to read calibration profile: {}",
-            profile_path.display()
-        )
-    })?;
-    info!(profile = %profile.name, delta = profile.energy_threshold_delta, "loaded calibration profile");
-    Ok(Some(profile.energy_threshold_delta))
-}
-
-fn load_tag_rules(profile_path: Option<&std::path::Path>) -> Result<Option<TagRules>> {
-    let Some(profile_path) = profile_path else {
-        return Ok(None);
-    };
-    let profile: CalibrationProfile = read_json(profile_path).with_context(|| {
-        format!(
-            "failed to read calibration profile: {}",
-            profile_path.display()
-        )
-    })?;
-    let rules = profile
-        .tag_thresholds
-        .as_ref()
-        .map(TagRules::from_thresholds);
-    if let Some(rules) = &rules {
-        info!(
-            profile = %profile.name,
-            ambience_max_rms = rules.ambience_max_rms,
-            impact_min_rms = rules.impact_min_rms,
-            min_centroid_hz = rules.min_centroid_hz,
-            "loaded tag rules from calibration profile"
-        );
-    } else {
-        info!(profile = %profile.name, "profile has no tag thresholds, using defaults");
-    }
-    Ok(rules)
-}
-
-fn load_merge_options(
-    config_path: Option<&std::path::Path>,
-    min_gap_override: Option<u32>,
-    strategy_override: Option<String>,
-) -> Result<MergeOptions> {
-    let cfg = if let Some(path) = config_path {
-        let data = std::fs::read_to_string(path).context("failed to read config file")?;
-        let analysis_cfg: crate::config::AnalysisConfig =
-            serde_json::from_str(&data).context("failed to parse config file")?;
-        analysis_cfg.merge_options.unwrap_or_default()
-    } else {
-        MergeOptions::default()
-    };
-
-    let mut opts = cfg;
-    if let Some(min_gap) = min_gap_override {
-        opts.min_gap_to_merge = min_gap;
-    }
-    if let Some(strategy) = strategy_override {
-        opts.merge_strategy = strategy;
-    }
-
-    validate_merge_options(&opts)?;
-    Ok(opts)
-}
-
-fn validate_merge_options(opts: &MergeOptions) -> Result<()> {
-    if !VALID_MERGE_STRATEGIES.contains(&opts.merge_strategy.as_str()) {
-        bail!(
-            "invalid merge_strategy: must be one of {}",
-            VALID_MERGE_STRATEGIES.join(", ")
-        );
-    }
-    if opts.min_gap_to_merge == 0 {
-        bail!("invalid min_gap_to_merge: must be > 0");
-    }
-    Ok(())
-}
-
 fn init_logging() {
     let filter =
         EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::from(Level::INFO.as_str()));
@@ -712,337 +614,4 @@ fn init_logging() {
         .with_target(false)
         .compact()
         .init();
-}
-
-fn open_in_browser(path: &std::path::Path) -> Result<()> {
-    let absolute = if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        std::env::current_dir()?.join(path)
-    };
-    let path_str = absolute.to_string_lossy().to_string();
-
-    if is_wsl() {
-        if try_open("wslview", &[&path_str])? {
-            info!(path = %absolute.display(), opener = "wslview", "opened review output in browser");
-            return Ok(());
-        }
-
-        if let Some(win_path) = wsl_to_windows_path(&path_str)? {
-            if try_open("cmd.exe", &["/C", "start", "", &win_path])? {
-                info!(path = %absolute.display(), opener = "cmd.exe/start", "opened review output in browser");
-                return Ok(());
-            }
-            if try_open(
-                "powershell.exe",
-                &["-NoProfile", "-Command", "Start-Process", &win_path],
-            )? {
-                info!(path = %absolute.display(), opener = "powershell.exe", "opened review output in browser");
-                return Ok(());
-            }
-        }
-    }
-
-    if cfg!(target_os = "macos") {
-        if try_open("open", &[&path_str])? {
-            info!(path = %absolute.display(), opener = "open", "opened review output in browser");
-            return Ok(());
-        }
-    } else if cfg!(target_os = "windows") {
-        if try_open("cmd", &["/C", "start", "", &path_str])? {
-            info!(path = %absolute.display(), opener = "cmd/start", "opened review output in browser");
-            return Ok(());
-        }
-        if try_open(
-            "powershell",
-            &["-NoProfile", "-Command", "Start-Process", &path_str],
-        )? {
-            info!(path = %absolute.display(), opener = "powershell", "opened review output in browser");
-            return Ok(());
-        }
-    } else {
-        let file_url = format!("file://{}", absolute.display());
-
-        if let Some(default_browser) = linux_default_browser_command()? {
-            if try_open(&default_browser, &["--new-window", &file_url])?
-                || try_open(&default_browser, &[&file_url])?
-            {
-                info!(
-                    path = %absolute.display(),
-                    opener = %default_browser,
-                    "opened review output in browser"
-                );
-                return Ok(());
-            }
-        }
-
-        if let Some(browser_env) = std::env::var_os("BROWSER") {
-            let browser_env = browser_env.to_string_lossy().to_string();
-            for candidate in browser_env.split(':').filter(|c| !c.is_empty()) {
-                if try_open(candidate, &[&file_url])? {
-                    info!(path = %absolute.display(), opener = %candidate, "opened review output in browser");
-                    return Ok(());
-                }
-            }
-        }
-
-        for candidate in ["google-chrome", "chromium-browser", "chromium", "firefox"] {
-            if try_open(candidate, &["--new-window", &file_url])?
-                || try_open(candidate, &[&file_url])?
-            {
-                info!(path = %absolute.display(), opener = %candidate, "opened review output in browser");
-                return Ok(());
-            }
-        }
-
-        if try_open("xdg-open", &[&path_str])? {
-            info!(path = %absolute.display(), opener = "xdg-open", "opened review output in browser");
-            return Ok(());
-        }
-        if try_open("gio", &["open", &path_str])? {
-            info!(path = %absolute.display(), opener = "gio open", "opened review output in browser");
-            return Ok(());
-        }
-    }
-
-    bail!(
-        "could not auto-open browser for {}; open it manually",
-        absolute.display()
-    )
-}
-
-fn linux_default_browser_command() -> Result<Option<String>> {
-    let output = match std::process::Command::new("xdg-settings")
-        .args(["get", "default-web-browser"])
-        .output()
-    {
-        Ok(output) if output.status.success() => output,
-        Ok(_) => return Ok(None),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-        Err(err) => return Err(err).context("failed running xdg-settings"),
-    };
-
-    let desktop = String::from_utf8_lossy(&output.stdout).trim().to_string();
-    if desktop.is_empty() {
-        return Ok(None);
-    }
-
-    let command = desktop.strip_suffix(".desktop").unwrap_or(&desktop).trim();
-    if command.is_empty() {
-        Ok(None)
-    } else {
-        Ok(Some(command.to_string()))
-    }
-}
-
-fn try_open(program: &str, args: &[&str]) -> Result<bool> {
-    match std::process::Command::new(program).args(args).status() {
-        Ok(status) => Ok(status.success()),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => Ok(false),
-        Err(err) => Err(err).with_context(|| format!("failed to run browser opener: {program}")),
-    }
-}
-
-fn wsl_to_windows_path(path: &str) -> Result<Option<String>> {
-    match std::process::Command::new("wslpath")
-        .args(["-w", path])
-        .output()
-    {
-        Ok(output) if output.status.success() => {
-            let win = String::from_utf8_lossy(&output.stdout).trim().to_string();
-            if win.is_empty() {
-                Ok(None)
-            } else {
-                Ok(Some(win))
-            }
-        }
-        Ok(_) => Ok(None),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => Ok(None),
-        Err(err) => Err(err).context("failed running wslpath"),
-    }
-}
-
-fn is_wsl() -> bool {
-    if std::env::var_os("WSL_DISTRO_NAME").is_some() || std::env::var_os("WSL_INTEROP").is_some() {
-        return true;
-    }
-    if let Ok(version) = std::fs::read_to_string("/proc/version") {
-        let lower = version.to_ascii_lowercase();
-        return lower.contains("microsoft") || lower.contains("wsl");
-    }
-    false
-}
-
-fn get_calibration_dir() -> Result<std::path::PathBuf> {
-    let base = if cfg!(target_os = "windows") {
-        std::env::var("APPDATA")?
-    } else if cfg!(target_os = "macos") {
-        std::path::PathBuf::from(std::env::var("HOME")?)
-            .join("Library/Application Support")
-            .to_string_lossy()
-            .to_string()
-    } else {
-        std::env::var("XDG_CONFIG_HOME")
-            .or_else(|_| std::env::var("HOME").map(|h| format!("{h}/.config")))
-            .map_err(|_| anyhow::anyhow!("Neither XDG_CONFIG_HOME nor HOME set"))?
-    };
-    Ok(std::path::PathBuf::from(base).join("do-movie-radio-play/profiles"))
-}
-
-fn merge_nonvoice_segments(
-    timeline: &crate::types::TimelineOutput,
-    options: &MergeOptions,
-    verified_keys: Option<&std::collections::HashSet<(u64, u64)>>,
-) -> crate::types::TimelineOutput {
-    let non_voice_segments: Vec<_> = timeline
-        .segments
-        .iter()
-        .filter(|s| {
-            if s.kind != crate::types::SegmentKind::NonVoice {
-                return false;
-            }
-            if let Some(keys) = verified_keys {
-                keys.contains(&(s.start_ms, s.end_ms))
-            } else {
-                true
-            }
-        })
-        .collect();
-
-    if non_voice_segments.is_empty() {
-        return crate::types::TimelineOutput {
-            file: timeline.file.clone(),
-            analysis_sample_rate: timeline.analysis_sample_rate,
-            frame_ms: timeline.frame_ms,
-            segments: vec![],
-        };
-    }
-
-    let merged = match options.merge_strategy.as_str() {
-        "all" => merge_all(non_voice_segments),
-        "longest" => merge_by_gap_threshold(non_voice_segments, options.min_gap_to_merge as u64),
-        "sparse" => merge_sparse_segments(non_voice_segments, options.min_gap_to_merge as u64),
-        _ => merge_all(non_voice_segments),
-    };
-
-    crate::types::TimelineOutput {
-        file: timeline.file.clone(),
-        analysis_sample_rate: timeline.analysis_sample_rate,
-        frame_ms: timeline.frame_ms,
-        segments: merged,
-    }
-}
-
-fn merge_all(segments: Vec<&crate::types::Segment>) -> Vec<crate::types::Segment> {
-    if segments.is_empty() {
-        return vec![];
-    }
-
-    let first_start = segments.first().map(|s| s.start_ms).unwrap_or(0);
-    let last_end = segments.last().map(|s| s.end_ms).unwrap_or(0);
-
-    let avg_confidence: f32 =
-        segments.iter().map(|s| s.confidence).sum::<f32>() / segments.len() as f32;
-
-    let all_tags: Vec<String> = segments.iter().flat_map(|s| s.tags.clone()).collect();
-
-    vec![crate::types::Segment {
-        start_ms: first_start,
-        end_ms: last_end,
-        kind: crate::types::SegmentKind::NonVoice,
-        confidence: avg_confidence,
-        tags: all_tags,
-        prompt: None,
-    }]
-}
-
-fn merge_by_gap_threshold(
-    segments: Vec<&crate::types::Segment>,
-    min_gap_ms: u64,
-) -> Vec<crate::types::Segment> {
-    let Some(first) = segments.first() else {
-        return vec![];
-    };
-
-    let mut merged = Vec::new();
-    let mut current_start = first.start_ms;
-    let mut current_end = first.end_ms;
-    let mut current_confidence = first.confidence;
-    let mut current_tags: Vec<String> = first.tags.clone();
-
-    for segment in segments.iter().skip(1) {
-        let gap = segment.start_ms - current_end;
-        if gap >= min_gap_ms {
-            merged.push(crate::types::Segment {
-                start_ms: current_start,
-                end_ms: current_end,
-                kind: crate::types::SegmentKind::NonVoice,
-                confidence: current_confidence,
-                tags: std::mem::take(&mut current_tags),
-                prompt: None,
-            });
-            current_start = segment.start_ms;
-            current_confidence = segment.confidence;
-            current_tags = segment.tags.clone();
-        }
-        current_end = segment.end_ms;
-    }
-
-    merged.push(crate::types::Segment {
-        start_ms: current_start,
-        end_ms: current_end,
-        kind: crate::types::SegmentKind::NonVoice,
-        confidence: current_confidence,
-        tags: current_tags,
-        prompt: None,
-    });
-
-    merged
-}
-
-fn merge_sparse_segments(
-    segments: Vec<&crate::types::Segment>,
-    min_gap_ms: u64,
-) -> Vec<crate::types::Segment> {
-    let Some(first) = segments.first() else {
-        return vec![];
-    };
-
-    let mut merged = Vec::new();
-    let mut current_start = first.start_ms;
-    let mut current_end = first.end_ms;
-    let mut current_confidence = first.confidence;
-    let mut current_tags: Vec<String> = first.tags.clone();
-
-    for segment in segments.iter().skip(1) {
-        let gap = segment.start_ms - current_end;
-        if gap >= min_gap_ms {
-            merged.push(crate::types::Segment {
-                start_ms: current_start,
-                end_ms: current_end,
-                kind: crate::types::SegmentKind::NonVoice,
-                confidence: current_confidence,
-                tags: std::mem::take(&mut current_tags),
-                prompt: None,
-            });
-            current_start = segment.start_ms;
-            current_confidence = segment.confidence;
-            current_tags = segment.tags.clone();
-        } else {
-            current_confidence = (current_confidence + segment.confidence) / 2.0;
-            current_tags.extend(segment.tags.clone());
-        }
-        current_end = segment.end_ms;
-    }
-
-    merged.push(crate::types::Segment {
-        start_ms: current_start,
-        end_ms: current_end,
-        kind: crate::types::SegmentKind::NonVoice,
-        confidence: current_confidence,
-        tags: current_tags,
-        prompt: None,
-    });
-
-    merged
 }
